+++ R/globals.R
@@ -1,6 +1,6 @@
 utils::globalVariables(c("RMSE", "adjR", "band_status", "choice", "control",
                          "controlID", "control_coefs", "estimate", "estimate_FE",
-                         "feols", "fixed", "included", "index", "is_fe",
+                         "included", "index", "is_fe",
                          "is_residual", "label", "lower", "n_obs", "observed",
                          "outside", "p", "percent", "ribbon.group", "se",
                          "se_type", "sig", "sig.level", "statistic", "term",